BUG#24848125: COLLECTION NAMES WITH DIFFERENT CASING ARE HIDDEN ON UBUNTU 14.04
Description:
X Plugin hides collections if the names have any letter in
a different case; for example, flags or Flags or flagS where the OS
is Linux. Where the OS is Windows, creating flags, Flags, and flagS
is not permitted.

Solution:
The 'list_objects' admin command SQL query has been improved to such
form to take into account a size of letters in names of tables in
a process of collecting them.

Reviewed-by: Lukasz Kotula <lukasz.kotula@oracle.com>
RB: 15173