Bug #25384527 DD TABLE SPATIAL_REFERENCE_SYSTEMS OVERSHOOTS THE INDEX LIMIT FOR 4K PAGE SIZE
Problem: The column st_spatial_reference_systems.name is defined as a
256 character long string in charset "utf8" with a unique index. If
run with innodb_page_size=4k, the string expanded to 3 bytes per
character exceeds the maximum index key length.

Fix: Reduce the length of the column to 80 characters. This is the
current limit for SRS names in the EPSG Dataset, which is the source
of the default set of SRSs.

Change-Id: I32e0a89de343ae3483a86f65289cc3df120b684f