BUG#24751177: ASSERTION `STRLEN(DB_NAME) <= (64*3) && STRLEN(TABLE_NAME) <= (64*3)' FAILED.
EXPLAIN/DESCRIBE of a table where the database name or the table
name is beyond the maximum supported length, an assert is triggered
in the debug build instead of reporting an error. In case of non
debug build, appropriate error is displayed.

During EXPLAIN/DESCRIBE, the temporary tables are processed
first followed by the regular tables. So we look up for the
temporary tables in the THD's temporary table list first. But
validation of database name and table name is done later in the
add_table_to_list(). Hence assert conditions to check length of
the table or db name while performing the look up in the temporary
table list failed.

In the current code, schema table implementation is used if
table is a temporary table or else new information schema
implementation(introduced by WL6599) is used. But performing a
look up in temporary table list before validating db and table
name looks incorrect.

As part of this patch, refactored code to validate db and table
name before performing table look up in temporary table list to
fix this issue.

Code refactoring involves following changes,

1) SHOW COLUMNS/SHOW INDEX implementation is moved to the new
   style,
   1.1 Parse_tree_node is introduced for the SHOW COLUMNS/SHOW
       INDEX statements.

   1.2 The decision of choosing schema table or new information
       schema implementation is moved to the contextualization
       stage.
        1.2.1 At this stage, table used in statement is validated
              first in add_table_to_list() and then lookup
              in temporary table list is performed.

   1.3 Sql_cmd_show is implemented to execute action for these
       statements.

2)  mysql_test_show() is cleaned up to remove the dead code.