Bug#24732452 - CA NOT VERIFIED WITH VERIFY_CA
DESCRIPTION
===========
While establishing connections from the client to the server
with SSL capability, connections succeed even if ssl-mode is
VERIFY_CA or VERIFY_IDENTITY and the client does not
provide CA certificate.

ANALYSIS
========
If ssl-mode is VERIFY_CA or VERIFY_IDENTITY, it must
either verify the server certificate against the
specified CA or fail the connection. Without providing
the CA or its path, the default CA is used and the
connection is still established.

FIX
===
We strictly enforce that the CA certifcate or its
path is provided if ssl-mode is one of the above two.
Otherwise, we throw an appropriate error message
and abort the connection.