BUG#24686908 INNODB SHOULD HAVE A MECHANISM TO CHECK IF ALL PFS KEYS ARE REGISTERED WITH PFS
Issue:
======
For the mutexes and rwlocks to be instrumented their respective PFS keys
must be registered with the PFS. It's possible that this registration is
missed and as a result the respective mutexes/rwlocks are not instrumented.
So it helps if there's a way to check if all the defined PFS keys are
registered with PFS.

Fix:
====
Figure out the number of PFS keys defined and number of PFS keys registered
with PFS and abort startup in debug build if the count doesn't match.

RB: 14913
Reviewed-by: Jimmy Yang <Jimmy.Yang@oracle.com>
Reviewed-by: Satya Bodapati <satya.bodapati@oracle.com>