Bug #25407875: GENERALIZE VARIABLE-LENGTH SORT KEY CODE
Currently, only JSON fields can be variable-length in sort keys, and a number
of details about this format are hard-coded to JSON and encoded in the JSON
sort keys function. We should make this more general, so that we can also
sort e.g. VARCHAR fields without padding them to the maximum length. (This is
especially important with utf8mb4 as default, where key lengths explode.)

In particular, we move the logic for writing field lengths out of
Field_json::make_sort_key() and into filesort itself. We also allow make_sort_key()
to return a length, and if sort_key_is_varlen() (a new member functions) returns
true, this length is allowed to be shorter than the full length of the field.

Similarly, the JSON sort key function no longer knows the details about where
the NULL indicator byte is supposed to be; it simply gets in a pointer to it
instead.

Change-Id: I581d5b8a04b9a17663eebc3de59d24aabe24df6e