Bug#25302901 NDB_PRINT_BACKUP_FILE SEGMENTATION FAULT WITH A TABLE WITH 500 COL
The ndb_print_backup_file tool reads table descriptors from the
.ctl file. While reading a table descriptor from the .ctl file
into the buffer, a buffer overflow occurs because the table
descriptor is too large to be accommodated within the buffer.
The buffer overflow corrupts the output stream object. The
corrupted output stream object is then used to print the table
descriptor, resulting in a segmentation fault.

Fix Description: Increased the buffer size so that the largest
possible table descriptor can be accommodated without overflow.