BUG#25408557 - PPOLL(2) DOES NOT ACCEPT NEGATIVE TIMEOUT VALUE
When a negative timeout value is passed to vio_io_wait, we pass
a negative value in timespec value to ppoll. This causes the
connection to idle and abort. Negative timeout value to
vio_io_wait means the call should block indefinitely
if no I/O event is available.

The fix is to pass null value to timespec argument of ppoll
to block indefinitely.