Bug#22705935 NDBAPI : SENDSIGNAL FLUSH OPTIMISATION ISSUES
Follow up patch to fix regression observed in the ATR
tests:

- testNdbApi -n ApiFailReqBehaviour T1
- testNdbApi -n CheckDisconnectCommit T1
- testNdbApi -n CheckDisconnectComplete T1
- testNdbApi -n FragmentedApiFailure T1
- ... and possibly more

The original patch removed the ::flush_send_signal()
from TransporterFacade::reportConnected / ::reportDisconnected()
after it had returned from ClusterMgr::reportConnected
or ::reportDisconnected().

This was done as flushing is now done by by the poll owner
when it 'finished poll'. However, that is probably too late
for the ::reportDisconnected() as we want the receiver to
take immediate actions on this - Possibly we even reset
the send buffers as part of disconnecting, such that the
unflushed signals are lost before they are sent, (Didn't
really investigate that though)

This patch reintroduce flushing for ClusterMgr::reportConnected()
and ::reportDisconnected() by changing their signal sending
to use the flushing ::safe_sendSignal() instead of
::safe_noflush_sendSignal(). This reintroduce the
immediate flushing of signals sent as part of a
connect/disconnect.