BUG#25200952 - TRANSACTIONS WITH DML WITH SUBQUERIES ON I_S TABLES CAN BLOCK CONCURRENT DDL
Problem:
--------
1. I_S queries are blocking concurrent DDLs
2. Two concurrent DDL statements which modify DD tables also do some reads
   from them explicitly and implicitly (as part of FK checks) acquiring
   gap locks in the process. As result deadlock occurs as these
   gap locks interact with inserts done by these statements.
3. GAP locks that are taken part of 'read for update' will block concurrent inserts
4. Next-key locks that are taken when inserting in unique secondary indexes block
   concurrent insert into DD tables.

Fix:
----
For 1)

These are pure reads as in SELECTS (and not read done to do update)

Use non-locking consistent reads for DD tables. This is applicable
not only for I_S queries but for any reads on DD tables.

For 2)
Prevent Gap locks that are taken as part of FK contrainst checks.

For 3)
We dont need Serialize/RR reads for DD tables, so GAP locks are not necessary
for DD tables.

For 4)
Use record locks instead of Next-Key locks. This will prevent blocking
the range and allow insert into unique secondary indexes. MDL on DD Table objects
prevents duplicates from being inserted.

Reviewed-By: Jimmy Yang <jimmy.yang@oracle.com>
RB: 15154