BUG#19641482 RPL.RPL_MTS_LOGICAL_CLOCK_RECOVERY FAILS WTH "SERVER FAILED              TO DISSAPPEAR"
Problem
-------

The test case is failing to make the slave server to "disappear".

Analysis
--------

The "crash_in_a_worker" debug sync point is relying on the fact that the
workload will be parallelized and reach MTS worker #2, but on slow
systems the parallelization will not happen and the server will fail to
"disappear".

Fix
---

Ensure that the workload will be distributed by at all the workers
even on slow systems.