Bug#24007040 - MYSQLDUMPSLOW ON MYSQL 5.7 DOESN'T FOLLOW                DATETIME FORMAT CHANGE
DESCRIPTION:
============
On MySQL 5.7, make_iso8601_timestamp() is used to build
timestamp string. This format is different from the
timestamp format of the previous versions. As a result,
mysqldumpslow cannot identify each entry and fails to
summarize the queries in slow-query-log.

ANALYSIS:
=========
The script's earlier regex pattern was not able to parse
the iso8601 timestamp format. As per the iso8601 timestamp
format, the timestamp can either be in the form of
YYYY-MM-DDTHH:MM:SS.SZ or
YYYY-MM-DDTHH:MM:SS.S[+-]HH:MM

FIX
====
Change the regex pattern for the timestamp according to the
iso8601 timestamp pattern. mysqldumpslow script should then
be able to summarize the queries correctly.