WL#9494 Implement INFORMATION_SCHEMA system views for ROUTINES/TRIGGERS/EVENTS
Post WL#9494 push fix:

In a very busy machines the test main.events_scheduling seem
to sporadically fail. The reason of failure is that querying
INFORMATION_SCHEMA.EVENTS sees the status of a event as
ENABLED instead of DISABLED. To make the test behave
consistent, this patch changes the wait condition to wait
until the status turn to DISABLED for events, instead of
waiting for ENABLED. This seem to avoid sporadic test
failure.

Although this failure is seen after WL#9494 is pushed, there
is no obvious reason to relate to. IMO, we uncovered this
failure now, may be because WL#9494 makes I_S.EVENTS
executes much faster than earlier.

Note: The failure could be reproduced by following command
      without the patch.

         ./mtr --ps-protocol
               --repeat=500
               --parallel=15
               main.events_scheduling{,,,}
               events_1{,,,,,}