Bug #23710632:  AFTER UPGRADING TO 5.7.13 LOGS   MANY [WARNING] EVENT 'MYSQL_AUDIT_CONNECT
Currently the information about the error used to try to abort the
audit command or the COM command is lost under a static message.

This makes it hard to debug the "blocked" messages in the error log.
The messages now print the underlying error too so that causes for it
can be prevented.