WL#6599 - New Data Dictionary and I_S integration.
Post-push fix: Sporadic failure of main.show_check test.

The test case for Bug#28808 is written to test behavior of
'log_queries_not_using_indexes' variable based on test that uses
INFORMATION_SCHEMA.TABLES. The expectation is that it should not
use index. But, after WL#6599, INFORMATION_SCHEMA.TABLES is a
system view which a JOIN over several DD tables and there is
possibility of optimizer opting to use a index. This leads
optimizer to set or unset SERVER_QUERY_NO_INDEX_USED as
thd->server_status sporadically. And this causes
log_slow_applicable() function to ignore logging the the query
when SERVER_QUERY_NO_INDEX_USED is not set to thd->server_status.

In order to fix the issue, the test case is modified to use a
user table 'tab1' (as described in bug#28808) instead of
INFORMATION_SCHEMA.TABLES. This avoid the sporadic failure.