Bug #25469683: REMOVE DESC/REVERSE OPTIONS TO WEIGHT_STRING
strnxfrm has a large set of flags that are not actually used by anything
_except_ our own debugging function weight_string(). Since this function is
not meant for end users (only for people developing collations), we can
remove some complexity.

The actual concept of reversing some levels remains valid, but we currently
have no collations doing it, and if they are to do it, it should be a property
of what the collation's strnxfrm function does, not something the strnxfrm
caller decides through a flag.

Change-Id: I65c56dba05114a9473ebac11ba07a79e0f190666