Bug#24666177
PARENT TABLE'S HANDLER NOT UPDATED AFTER ADDING A FK DEPENDENCY

When a foreign key is added or dropped from a table using the alter
table query, the child's metadata is reloaded in the mysqld handler. But
the parent table's metadata is not reloaded. Due to this the fk
information in parent's handler becomes obsolete. When a subsequent
query comes in that depends on this particular foreign key information,
it might return bad results. Similar problem happens also after a
foreign key drop.

This patch fixes that by flushing out the cached parent table definition
after a successful foreign key ddl query.