WL #10354: Make Unicode 9.0.0 and newer collations NO PAD (update CHAR sort keys)
Prepare CHAR(n) for NO PAD collations; we need to make sort keys consistent
with val_str(), so that their space-stripping behavior is the same.
This wasn't done before because only BINARY was a NO PAD collation, and it
had special-cased behavior (it doesn't pad with spaces, it pads with zero bytes).

Change-Id: I5039db0a2e36526120554c87604714208f22da1c