WL#7318 Delayed Replication: GTID based and relative to immediate master commit
Delayed replication is now applied per transaction (instead
of per event) and is computed relative to the
immediate_commit_timestamp (introduced in WL#7319) if both
the immediate master and current server (slave) support
immediate_commit_timestamp.
If at least one of the servers does not support this
timestamp, the delay will be computed using the old
implementation.