BUG#25256910: NODE RECOVERY FAILS WITH ERROR_CODE: 1062; HANDLER ERROR HA_ERR_FOUND_DUPP_KEY
On Single Primary Mode Group Replication certification process is disabled,
only one server accepts writes so there aren't conflicts. Exist one exception
for when a new primary server is being elected.

This causes a problem for when a new server is entering the group, it doesn't
know if certification is enable or not in the group as a consequence of the
primary election mechanism. Hence, queries being rejected in other member can
be accepted and fail on the new joiner.

This patch solves this problem sending on view change, when a member join or
leaves, the state of conflict detection of the members. When the member enters
it will receive the state of conflict detection of all members and activate it
if one of the members has conflict detection active.