Bug#25522709 - MAIN.MYSQLHOTCOPY_ARCHIVE AND                MAIN.MYSQLHOTCOPY_MYISAM TESTS FAILING                IN PB2 IN 5.6
DESCRIPTION:
============
main.mysqlhotcopy_archive and main.mysqlhotcopy_myisam tests
are failing in PB2 weekly-5.6 on the freebsd platforms.

ANALYSIS:
========
This is caused by an explicit shebang in
checkDBI_DBD-mysql.pl and in mysqlhotcopy.sh using
#!/usr/bin/perl which is not correct on a standard install
of FreeBSD 11. This results in a failure of the script
execution.

FIX:
====
The fix is to use a portable shebang -> #!/usr/bin/env perl
instead of #!/usr/bin/perl.