Bug#25516881: JSON_UNQUOTE RETURNS WRONG RESULT WHEN INPUT IS UTF-16
The JSON_UNQUOTE function did not unquote the string if it was UTF-16
encoded.

This happened because the function had some code to check if the
string started and ended with double-quotes, which assumed that the
input was UTF-8 encoded. This made it think that UTF-16 strings always
started with '\0', so it never treated them as quoted strings.

Fix: Convert the input string to utf8mb4, if needed, before attempting
to analyze the contents of the string.

Change-Id: I1c3fd717cb14ea8fcd57d99f5e9f7a61da798dce