Fix ATR test 'testScan -n ScanDuringShrinkAndExpandBack' (Only seen failing in private sandbox)
The 'NdbDictionary::Column::ROW_COUNT' of the test table
is read back by setting up 'extraGetValues' in the
'ScanOptions'. This require the adress of an *Uint64*
to be suplied which the ROW_COUNT will be read into.

However, the test program incorrectly suplied an
Uint32 instead of an Uint64. Thus the 4 bytes following
the '&row_count' argument was also written into when
a row_count is returned.