Bug#25530204: THE RESULT OF SUM ON JSON_EXTRACT LOST THE DECIMAL PART
When using the SUM aggregate function on an expression that returns a
JSON value, the result does not have a fractional part.

The reason is that the Item_json_func class does not change
Item::decimals from its initial zero value. This is used to determine
how the result is presented, and when it is zero, the fractional part
is lost.

Fix: Set Item::decimals to NOT_FIXED_DEC (31) in
Item_json_func::resolve_type(). This is the same value that is used
for other string types.

For consistency, setting this member, and other members common to
expressions that return JSON values, is moved into a separate helper
function, Item::set_data_type_json(). Item_json and Item_sum_json are
also updated to using the helper function.

Item_sum_json redundantly set the data type to JSON both in its
constructors and in fix_fields(). The patch removes the code from
fix_fields() so that it is only set once. Similarly, make
Item_json_func set the data type in its constructors instead of in
resolve_type().

Changed test result:

Two test cases in json.json_agg get slightly different output because
the type of 1.0 * JSON_ARRAYAGG(a) changed from DOUBLE(18, 1) to
DOUBLE, and the type of SEC_TO_TIME(JSON_ARRAYAGG(a)) changed from
TIME(6) to TIME.

x.find_doc_proj shows metadata for some results where
"fractional_parts" is changed from 0 to 31 as expected.

Change-Id: I37a24afa463201059b7232583d60a5d2e6774fe0