Bug#25372805 PERFORMANCE SCHEMA 5.6.35 UNITTEST PFT-T FAILING
Unit test pfs-t is failing, with the following error printed:
  not ok 179 - no locker (no thread owner)

The root cause is the test case for sockets,
which was not updated to follow a change in implementation in the code,
as the behavior changed.

Now, somehow this test failure went unnoticed,
and after investigation it turns out that the unit tests
programs to not return the test exit_status() on exit,
masking test failures to the automated build scripts.

This fix adjusts the failing test case,
and fixes all performance schema unit tests
to properly return the test exit_status().