C++ conformance fixes (MSVC /permissive-)
We (the Microsoft C++ team) use the MySQL project as part of our "Real world code" tests.
I noticed a few issues in windows specific code when building MySQL with the MSVC compiler
in its conformance mode (/permissive-).  For more information on /permissive- see our blog
https://blogs.msdn.microsoft.com/vcblog/2016/11/16/permissive-switch/.

These changes are to address 3 different types of issues:

1) Use of qualified names in member declarations

    struct A {
        void A::f() { } // error C4596: illegal qualified name in member declaration
                        // remove redundant 'A::' to fix
    };

2) String literal type conversion

    char * str = "string"; // String literal loses 'const'
	                       // either change to 'const char*' or cast literal to 'char*'

I cast the value for shared_memory_base_name because adding 'const' to something like that
can have a big domino effect and I'm not sure I have the resourses to verify a change like that on
other systems.  You may want to make it a low priority 'todo' item for the future.  For now
it makes the current compiler behavior explicit.

3) Do not allow skipping an initializer with a goto

    void func(bool b)
    {
        if(b) goto END;
  
        int value = 0; //error C2362: initialization of 'value' is skipped by 'goto END'
              int array[10]; //Okay, not initialized.
  
        //... value used here
  
    END:
        return;
    }

One potential fix is to limit the scope of the initialized variable so it ends before the label.
See our blog for more potential fixes.