Bug #20607730 CLUSTER FAILED TO SHUTDOWN GRACEFULLY....
c_waitGCPMasterList is an array pool and its hard coded to 10. When ever
graceful shutdown is initiated master node gets WAIT_GCP_REQ signal equal
to the no of nodes configured. So, as it is hard coded to 10 when we have
a cluster with more than 10 nodes, due to lack of pool resource we always
get "Forced node shutdown completed. Caused by error 2305: 'Node lost
connection to other nodes and cannot form a un-partitioned cluster, please
investigate if there are error(s) on other node(s)(Arbitration error).
Temporary error, restart node'."

So, to avoid this have changed the pool size to the max no of nodes that
can be configured in the cluster.