BUG#25584734 SHOW BINLOG/RELAYLOG EVENTS BLOCKED WAITING FOR LOCK_LOG
Problem/Analysis
----------------

The access to the binary/relay log events by SHOW BINLOG|RELAYLOG EVENTS
blocked the log access by acquiring the log lock_log for the whole
"dump", but this was fixed in BUG#20928790.

After BUG#20928790 was fixed, SHOW BINLOG|RELAYLOG EVENTS statements
acquire the log's lock_log only for a brief amount of time to calculate
the end of the log file. This was necessary in MySQL 5.6 in order to
avoid to read "partial" events while the binary/relay log was being
written (by commits or by the receiver thread).

In MySQL 5.7, WL#5721 has implemented the binlog_end_pos, to allow the
dump thread to read from the binary log (even from the "hot" log file)
without the need of acquiring the log's lock_log. In this way, no dump
thread would block commits from happen.

In MySQL 8.0.1, WL#8599 made the applier threads to rely on
binlog_end_pos too.

Fix
---

Now, SHOW BINLOG|RELAYLOG EVENTS was modified to
rely on the binlog_end_pos for the "hot" log file. This made the
statements to not block commits or receiving events at all.