Bug #25588520: COLLATIONS USE CLDR_29 IN VARIABLE NAMES ALTHOUGH THEY ARE BASED ON CLDR 30
When we were implementing new UCA collations, they were based on UCA 8.0.0
and CLDR v29. Then we changed to latest UCA 9.0.0. We found that there is
no difference between CLDR v29 and v30 on the collations we have. So we
didn't change the variable names. But we agree that this really raises
confusion and decide to change them now.

Change-Id: I4e1cd93c489952224fd9e544e6b5d062ed898450